fix tests and more

+++ libs/blockscout-client/crate/src/models/reward.rs
@@ -18,8 +18,7 @@ use serde_aux::prelude::*;
 
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize, derive_new::new)]
 pub struct Reward {
-    // false true false true
     #[serde(deserialize_with = "deserialize_number_from_string")]
-    pub reward: i64,
+    pub reward: f64,
     pub r#type: String,
 }
